fix anonymous classes
