+++ kogito-codegen-modules/kogito-codegen-processes/src/main/resources/class-templates/usertask/RestResourceUserTaskQuarkusTemplate.java
@@ -50,6 +50,8 @@ import org.kie.kogito.process.impl.Sig;
 import org.kie.kogito.services.uow.UnitOfWorkExecutor;
 import org.kie.kogito.usertask.UserTaskInstanceNotFoundException;
 import org.kie.kogito.usertask.UserTaskService;
+import org.kie.kogito.usertask.impl.json.SimpleDeserializationProblemHandler;
+import org.kie.kogito.usertask.impl.json.SimplePolymorphicTypeValidator;
 import org.kie.kogito.usertask.view.UserTaskView;
 import org.kie.kogito.usertask.view.UserTaskTransitionView;
 
@@ -82,33 +84,9 @@ public class UserTasksResource {
     public void init() {
         mapper = objectMapper.copy();
         SimpleModule module = new SimpleModule();
-        mapper.addHandler(new DeserializationProblemHandler() {
-            @Override
-            public JavaType handleMissingTypeId(DeserializationContext ctxt, JavaType baseType, TypeIdResolver idResolver, String failureMsg) throws IOException {
-                return baseType;
-            }
-        });
+        mapper.addHandler(new SimpleDeserializationProblemHandler());
         mapper.registerModule(module);
-
-        PolymorphicTypeValidator validator = new PolymorphicTypeValidator() {
-
-            @Override
-            public Validity validateBaseType(MapperConfig<?> config, JavaType baseType) {
-                return Validity.ALLOWED;
-            }
-
-            @Override
-            public Validity validateSubClassName(MapperConfig<?> config, JavaType baseType, String subClassName) throws JsonMappingException {
-                return Validity.ALLOWED;
-            }
-
-            @Override
-            public Validity validateSubType(MapperConfig<?> config, JavaType baseType, JavaType subType) throws JsonMappingException {
-                return Validity.ALLOWED;
-            }
-            
-        };
-        mapper.activateDefaultTypingAsProperty(validator, DefaultTyping.NON_FINAL, "@type");
+        mapper.activateDefaultTypingAsProperty(new SimplePolymorphicTypeValidator(), DefaultTyping.NON_FINAL, "@type");
     }
 
     @GET
@@ -131,7 +109,7 @@ public class UserTasksResource {
     public UserTaskView transition(
             @PathParam("taskId") String taskId,
             @QueryParam("user") String user,
-            @QueryParam("group") List<String> groups, 
+            @QueryParam("group") List<String> groups,
             TransitionInfo transitionInfo) {
         return userTaskService.transition(taskId, transitionInfo.getTransitionId(), transitionInfo.getData(), IdentityProviders.of(user, groups)).orElseThrow(UserTaskInstanceNotFoundException::new);
     }
@@ -165,7 +143,7 @@ public class UserTasksResource {
             @PathParam("taskId") String taskId,
             @QueryParam("user") String user,
             @QueryParam("group") List<String> groups,
-            String body) throws Exception{
+            String body) throws Exception {
         Map<String, Object> data = mapper.readValue(body, Map.class);
         return userTaskService.setInputs(taskId, data, IdentityProviders.of(user, groups)).orElseThrow(UserTaskInstanceNotFoundException::new);
     }
